+++ Modules/Intent.Modules.AzureFunctions.Interop.EntityFrameworkCore/Decorators/DbContextSaveFunctionDecorator.cs
@@ -47,8 +47,7 @@ namespace Intent.Modules.AzureFunctions.Interop.EntityFrameworkCore.Decorators
                         .AddStatement("return new NotFoundObjectResult(new { Message = exception.Message });"));
             });
             
-            if (HttpEndpointModelFactory.GetEndpoint(_template.Model.InternalElement)?.Verb == HttpVerb.Get ||
-                template.Model.Mapping?.Element?.AsOperationModel() == null)
+            if (HttpEndpointModelFactory.GetEndpoint(_template.Model.InternalElement)?.Verb == HttpVerb.Get)
             {
                 return;
             }