+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/ListedServiceFunc.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using AzureFunctions.TestApplication.Application.Interfaces;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
+using AzureFunctions.TestApplication.Domain.Common.Interfaces;
 using FluentValidation;
 using Intent.RoslynWeaver.Attributes;
 using Microsoft.AspNetCore.Http;
@@ -25,10 +26,12 @@ namespace AzureFunctions.TestApplication.Api
     public class ListedServiceFunc
     {
         private readonly IListedUnlistedServicesService _appService;
+        private readonly IUnitOfWork _unitOfWork;
 
-        public ListedServiceFunc(IListedUnlistedServicesService appService)
+        public ListedServiceFunc(IListedUnlistedServicesService appService, IUnitOfWork unitOfWork)
         {
             _appService = appService ?? throw new ArgumentNullException(nameof(appService));
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         }
 
         [FunctionName("ListedServiceFunc")]
@@ -43,6 +46,7 @@ namespace AzureFunctions.TestApplication.Api
             {
                 string param = req.Query["param"];
                 await _appService.ListedServiceFunc(param);
+                await _unitOfWork.SaveChangesAsync();
                 return new CreatedResult(string.Empty, null);
             }
             catch (NotFoundException exception)