+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/UnMappedAzureFunction.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using AzureFunctions.TestApplication.Application.SampleDomains;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
+using AzureFunctions.TestApplication.Domain.Common.Interfaces;
 using Intent.RoslynWeaver.Attributes;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -22,8 +23,10 @@ namespace AzureFunctions.TestApplication.Api
 {
     public class UnMappedAzureFunction
     {
-        public UnMappedAzureFunction()
+        private readonly IUnitOfWork _unitOfWork;
+        public UnMappedAzureFunction(IUnitOfWork unitOfWork)
         {
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         }
 
         [FunctionName("UnMappedAzureFunction")]