+++ Modules/Intent.Modules.AzureFunctions.Dispatch.MediatR/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.5
+
+- Added the necessary File Builder metadata in order to indicate which line of code has to do with service dispatching. This is important for other modules that want to hook-into that space.
+
 ### Version 1.0.3
 
 - Fix: Azure Route info for `Http Tigger` on `ServiceOperations`.